Move the map implementation out of pages
