+++ src/doc-containers/map-implementation/layer-manager/layers/example.json
