+++ src/doc-containers/map-implementation/layer-manager/layers/index.ts
@@ -1,0 +1,1 @@
+export { default } from './example-layer';
