+++ src/doc-containers/map-implementation/layer-manager/types.d.ts
