+++ src/pages/docs/components.mdx
@@ -6,6 +6,7 @@ import DocsLayout from 'layouts/docs'
 First try to find the component you need in the [shadcn library](https://ui.shadcn.com/docs). This is our reference for new components.
 If you can't find it there, use one of the components availiable on the components folder:
 
+- [Icon](/docs/components/icon)
 - [Map](/docs/components/map)
 
 export default ({ children }) => <DocsLayout>{children}</DocsLayout>
