+++ src/pages/docs/components/icon/index.mdx
@@ -1,0 +1,9 @@
+import DocsLayout from 'layouts/docs';
+import FIT_BOUNDS_SVG from 'svgs/map/fit-bounds.svg?sprite';
+import Icon from 'components/icon';
+
+# Icon component
+
+<Icon icon={FIT_BOUNDS_SVG} />
+
+export default ({ children }) => <DocsLayout>{children}</DocsLayout>
