+++ src/pages/docs/components/map/index.mdx
@@ -1,5 +1,5 @@
 import DocsLayout from 'layouts/docs'
-import MapImplementation from './map-implementation';
+import MapImplementation from 'doc-containers/map-implementation';
 
 # Map component
