+++ tailwind.config.js
@@ -10,6 +10,8 @@ module.exports = {
   content: [
     './src/components/**/*.@(tsx|ts)',
     './src/containers/**/*.@(tsx|ts)',
+    // Remove this line if the documentation is not used
+    './src/doc-containers/**/*.@(tsx|ts)',
     './src/layouts/**/*.@(tsx|ts)',
     './src/pages/**/*.@(tsx|ts)',
   ],
